requirements.txt
playwright>=1.44.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.36
greenlet>=3.1.0
psycopg2-binary>=2.9.10
anthropic>=0.40.0
boto3>=1.34.0
python-dotenv>=1.0.0
tenacity>=8.3.0
